+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.1] - 2023-03-13
+
 ### Added
 
 - Added private function `Build-ADAuditTasksUser` for ADAuditTasksUser Class.